refactor(practice page): move logic to hook
